requirements.txt
python-telegram-bot==21.4
feedparser==6.0.11
python-dotenv==1.0.1
beautifulsoup4==4.12.3
html5lib==1.1
